+++ src/lib.rs
@@ -53,6 +53,12 @@ pub use crate::tables::gdef::{
     LigCaretList, LigGlyph, MarkGlyphSets,
 };
 pub use crate::tables::gpos::GposTable as GposView;
+pub use crate::tables::gpos::{
+    SinglePos, SinglePosIter, ValueFormat, ValueRecord, GPOS_LOOKUP_TYPE_CHAINED_CONTEXT,
+    GPOS_LOOKUP_TYPE_CONTEXT, GPOS_LOOKUP_TYPE_CURSIVE, GPOS_LOOKUP_TYPE_EXTENSION,
+    GPOS_LOOKUP_TYPE_MARK_TO_BASE, GPOS_LOOKUP_TYPE_MARK_TO_LIGATURE,
+    GPOS_LOOKUP_TYPE_MARK_TO_MARK, GPOS_LOOKUP_TYPE_PAIR, GPOS_LOOKUP_TYPE_SINGLE,
+};
 pub use crate::tables::gsub::GsubTable as GsubView;
 pub use crate::tables::gsub::{
     AlternateGlyphIter, AlternateSet, AlternateSubst, AlternateSubstIter, ExtensionSubst, Ligature,